requirements.txt
#This file is for deployment in Deta
fastapi
python-jose
passlib
pymongo
python-multipart
certifi